:cookie: Redesign the last column on a jira-board
If you processed a lot of tickets in short time the last column tends
to be overfilled by tickets. That creates an annoying long page scroll
to find open tickets which are not rated so high.

This change adds some CSS-Rules to redesign these tiles and make them
tiny and easy to handle. This causes a minimal drop of information but
this information should also be provided in the detailed view of the
ticket.