:bug: Correct the position of the translation menu
If the page was scrolled in any direction the calculation of the
mouse-position failed because the scrolling was added twice.

The scrolling was added on one hand by the mouse-position relative to
the document and on the other hand by the boundary-rectangle of the body
which will be moved relative to the document if the page is scrolled.

Therefore the calculation is now the following which handles almost
every CSS and scroling influence:
mouseposition - boundary-rectangle - scrolling